+++ Estimation/Merge/PoissonRealBarcodesMergeStrategy.h
@@ -25,7 +25,7 @@ namespace Merge
 		virtual void release() override;
 
 		virtual long get_best_merge_target(const CellsDataContainer &container, size_t base_cell_ind, const ul_list_t &neighbour_cells) const override;
-		virtual long get_max_merge_dist(long min_real_cb_dist) const override;
+		virtual unsigned get_max_merge_dist(unsigned min_real_cb_dist) const override;
 
 	public:
 		PoissonRealBarcodesMergeStrategy(const std::string &barcodes_filename, const boost::property_tree::ptree &config);
